[0.2] Add option --extends={parent} to make:class command
