+++ src/Console/stubs/class.parent.stub
@@ -1,0 +1,10 @@
+<?php
+
+namespace DummyNamespace;
+
+use DummyParentFullName;
+
+class DummyClass extends DummyParentName
+{
+    //
+}
